+++ lib/nfg_ui/components/traits/step.rb
@@ -3,9 +3,18 @@ module NfgUi
     module Traits
       # Step traits
       module Step
-        TRAITS = %i[visited].freeze
+        TRAITS = %i[visited active].freeze
 
         def visited_trait
+          options[:visited] = true
+        end
+
+        def active_trait
+          super
+          # Active is a transitory state for steps
+          # So when a step is active, it is automatically visited.
+          # This sets visited to true.
+          # This is also accounted for in the #active_trait for steps.
           options[:visited] = true
         end
       end